test: move assign action tests in its own file
