+++ mergify_engine/tests/functional/test_engine.py
@@ -164,27 +164,6 @@ expected alphabetic or numeric character, but found"""
             checks[0]["output"]["title"],
         )
 
-    def test_assign(self):
-        rules = {
-            "pull_request_rules": [
-                {
-                    "name": "assign",
-                    "conditions": [f"base={self.master_branch_name}"],
-                    "actions": {"assign": {"users": ["mergify-test1"]}},
-                }
-            ]
-        }
-
-        self.setup_repo(yaml.dump(rules))
-
-        p, _ = self.create_pr()
-
-        pulls = list(self.r_o_admin.get_pulls(base=self.master_branch_name))
-        self.assertEqual(1, len(pulls))
-        self.assertEqual(
-            sorted(["mergify-test1"]), sorted([l.login for l in pulls[0].assignees])
-        )
-
     def test_request_reviews_users(self):
         rules = {
             "pull_request_rules": [
